Add gzip support
Here I'm just adding gzip to the options and if it is enabled, then we
run the file through a gzip before sending it to the cache and the
user. I also store the encoding in the cache object so multiple
encoding types can be supported. Just incase deflate were to be added.